+++ src/main/java/com/insane96mcp/shieldsplus/world/item/enchantment/ShieldBashEnchantment.java
@@ -1,6 +1,7 @@
 package com.insane96mcp.shieldsplus.world.item.enchantment;
 
 import com.insane96mcp.shieldsplus.setup.SPEnchantments;
+import com.insane96mcp.shieldsplus.setup.Strings;
 import insane96mcp.insanelib.util.MCUtils;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.nbt.CompoundTag;
@@ -24,6 +25,8 @@ import java.util.List;
 import java.util.UUID;
 
 public class ShieldBashEnchantment extends Enchantment {
+
+	static UUID STEP_HEIGHT_ON_BASH = UUID.fromString("259f26e7-6914-4856-a01f-8b2295e6d244");
 
 	public ShieldBashEnchantment() {
 		super(Rarity.VERY_RARE, EnchantmentCategory.BREAKABLE, new EquipmentSlot[]{EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
@@ -57,41 +60,39 @@ public class ShieldBashEnchantment extends Enchantment {
 	}
 
 	public static void onTick(Player player) {
-		int shieldBash = EnchantmentHelper.getItemEnchantmentLevel(SPEnchantments.SHIELD_BASH.get(), player.getUseItem());
-		if (shieldBash <= 0)
-			return;
-
 		CompoundTag tag = player.getPersistentData();
-		if (player.isBlocking() && tag.getByte("shield_bashing") <= 0) {
-			if (tag.getByte("bash_timer") < 30) {
-				tag.putByte("bash_timer", (byte) (tag.getByte("bash_timer") + 1));
-				if (tag.getByte("bash_timer") >= 30)
+		int shieldBash = EnchantmentHelper.getItemEnchantmentLevel(SPEnchantments.SHIELD_BASH.get(), player.getUseItem());
+
+		if (tag.getByte(Strings.Tags.BASH_TIMER) < 0){
+			tag.putByte(Strings.Tags.BASH_TIMER, (byte) (tag.getByte(Strings.Tags.BASH_TIMER) + 1));
+		}
+
+		if (player.isBlocking() && tag.getByte(Strings.Tags.SHIELD_BASHING) <= 0 && shieldBash > 0) {
+			if (tag.getByte(Strings.Tags.BASH_TIMER) < 30) {
+				tag.putByte(Strings.Tags.BASH_TIMER, (byte) (tag.getByte(Strings.Tags.BASH_TIMER) + 1));
+				if (tag.getByte(Strings.Tags.BASH_TIMER) >= 30)
 					player.playSound(SoundEvents.SHIELD_BLOCK, 1f, 1.65f);
 			}
 			else if (player.isCrouching() && player.isOnGround()) {
-				//player.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SLOWDOWN, 10, 10, true, true, false));
-				tag.putByte("shield_bashing", (byte) 12);
-				MCUtils.applyModifier(player, ForgeMod.STEP_HEIGHT_ADDITION.get(), UUID.fromString("259f26e7-6914-4856-a01f-8b2295e6d244"), "Shield Bash step height", 0.6d, AttributeModifier.Operation.ADDITION, false);
+				tag.putByte(Strings.Tags.SHIELD_BASHING, (byte) 12);
+				MCUtils.applyModifier(player, ForgeMod.STEP_HEIGHT_ADDITION.get(), STEP_HEIGHT_ON_BASH, "Shield Bash step height", 0.6d, AttributeModifier.Operation.ADDITION, false);
 				float f = -Mth.sin(player.getYRot() * ((float) Math.PI / 180F));
 				float f1 = Mth.cos(player.getYRot() * ((float) Math.PI / 180F));
-				player.playSound(SoundEvents.SHIELD_BLOCK, 1f, 2.0f);
+				player.playSound(SoundEvents.SHIELD_BLOCK, 1f, 1.3f);
 				player.setDeltaMovement(player.getDeltaMovement().add(f * 2.75d, 0.4d, f1 * 2.75d));
 				for (int i = 0; i < 50; i++) {
 					player.level.addParticle(ParticleTypes.CLOUD, player.getX() + Mth.nextDouble(player.getRandom(), -0.5d, 0.5d), player.getY() + Mth.nextDouble(player.getRandom(), -0.5d, 0.5d) + 0.9d, player.getZ() + Mth.nextDouble(player.getRandom(), -0.5d, 0.5d), 0.1, 0, 0.1);
 				}
-				tag.putByte("bash_timer", (byte) -30);
+				tag.putByte(Strings.Tags.BASH_TIMER, (byte) -30);
 			}
 		}
-		else if (tag.getByte("bash_timer") > 30){
-			tag.putByte("bash_timer", (byte) 0);
-		}
-		else {
-			tag.putByte("bash_timer", (byte) (tag.getByte("bash_timer") + 1));
+		else if (tag.getByte(Strings.Tags.BASH_TIMER) > 0) {
+			tag.putByte(Strings.Tags.BASH_TIMER, (byte) 0);
 		}
 
-		if (tag.getByte("shield_bashing") > 0) {
-			tag.putByte("shield_bashing", (byte) (tag.getByte("shield_bashing") - 1));
-			if (tag.getByte("shield_bashing") == 0) {
+		if (tag.getByte(Strings.Tags.SHIELD_BASHING) > 0) {
+			tag.putByte(Strings.Tags.SHIELD_BASHING, (byte) (tag.getByte(Strings.Tags.SHIELD_BASHING) - 1));
+			if (tag.getByte(Strings.Tags.SHIELD_BASHING) == 0) {
 				if (player.getAttribute(ForgeMod.STEP_HEIGHT_ADDITION.get()) != null)
 					//noinspection ConstantConditions
 					player.getAttribute(ForgeMod.STEP_HEIGHT_ADDITION.get()).removeModifier(UUID.fromString("259f26e7-6914-4856-a01f-8b2295e6d244"));
@@ -103,7 +104,7 @@ public class ShieldBashEnchantment extends Enchantment {
 	}
 
 	public static void damageAndKnockback(Player player) {
-		AABB aabb = player.getBoundingBox().inflate(1.25d, 0.5d, 1.25d);
+		AABB aabb = player.getBoundingBox().inflate(1.2d, 0.5d, 1.2d);
 		List<LivingEntity> entities = player.level.getEntitiesOfClass(LivingEntity.class, aabb, entity -> entity != player);
 		for (LivingEntity entity : entities) {
 			if (entity.isDeadOrDying() || entity.invulnerableTime > 10)